+++ python/marvin/web/lib/src/galaxy.js
@@ -32,9 +32,10 @@ class MapError extends Error {
 class Galaxy {
 
     // Constructor
-    constructor(plateifu, toggleon) {
+    constructor(plateifu, toggleon, redshift) {
         this.setPlateIfu(plateifu);
         this.toggleon = toggleon;
+        this.redshift = redshift;
         // main elements
         this.maindiv = $('#'+this.plateifu);
         this.metadiv = this.maindiv.find('#metadata');
@@ -53,8 +54,12 @@ class Galaxy {
         this.togglediv = $('#toggleinteract');
         this.toggleload = $('#toggle-load');
         this.togglediv.bootstrapToggle('off');
+        this.toggleframe = $('#toggleframe');
+        this.togglelines= $('#togglelines');
+        this.togglemask = $('#togglemask'); 
         // flag popover elements
         this.qualpop = $('#qualitypopover');
+        this.mapqualpop = $('#mapqualitypopover');
         this.targpops = $('.targpopovers');
         // maps elements
         this.dapmapsbut = $('#dapmapsbut');
@@ -92,6 +97,9 @@ class Galaxy {
         this.nsaplotbuttons.on('click', this, this.updateNSAPlot);
         //this.nsatable.on('page-change.bs.table', this, this.updateTableEvents);
         //this.nsatable.on('page-change.bs.table', this, this.updateTableEvents);
+        this.toggleframe.on('change', this, this.toggleWavelength); // this event fires when a user clicks the Toggle Obs/Rest Frame
+        this.togglelines.on('change', this, this.toggleLines); // this event fires when a user clicks the Toggle Lines
+        this.togglemask.on('change', this, this.toggleMask); // this event fires when a user clicks the Toggle Masks
 
     }
 
@@ -120,9 +128,16 @@ class Galaxy {
         }, 10);
     }
 
+    // Compute rest-frame wavelength
+    computeRestWave() {
+        let rest = this.setSpectrumAxisFormatter('rest', this.redshift);
+        this.obswave = this._spaxeldata.map(x => x[0]);
+        this.restwave = this.obswave.map(x => rest(x));
+        this.rest_spaxeldata = this._spaxeldata.map((x, i) => [this.restwave[i], x.slice(1)[0], x.slice(1)[1]]);
+    }
+
     // Initialize and Load a DyGraph spectrum
     loadSpaxel(spaxel, title) {
-        this._spaxeldata = spaxel;
         // this plugin renables dygraphs 1.1 behaviour of unzooming to specified valueRange 
         const doubleClickZoomOutPlugin = {
             activate: function (g) {
@@ -140,35 +155,149 @@ class Galaxy {
             }
         };
 
+        // run intial setup of spaxel data
+        this.setupSpaxel(spaxel);
+
+        // initialize Dygraph
         let labels = (spaxel[0].length == 3) ? ['Wavelength','Flux', 'Model Fit'] : ['Wavelength','Flux'];
-        this.webspec = new Dygraph(this.graphdiv[0],
-                  spaxel,
-                  {
-                    title: title,
-                    labels: labels,
-                    errorBars: true,  // TODO DyGraph shows 2-sigma error bars FIX THIS
-                    ylabel: 'Flux [10<sup>-17</sup> erg/cm<sup>2</sup>/s/Å]',
-                    xlabel: 'Observed Wavelength [Ångströms]',
-                    valueRange: [0,null],
-                    plugins: [doubleClickZoomOutPlugin],
-                    axes: {
-                        x: {
-                            axisLabelFormatter: this.setSpectrumAxisFormatter('obs')
-                        }
-                    }
-                });
+        let options = {
+            title: title,
+            labels: labels,
+            legend: 'always',
+            errorBars: true,  // TODO DyGraph shows 2-sigma error bars FIX THIS
+            ylabel: 'Flux [10<sup>-17</sup> erg/cm<sup>2</sup>/s/Å]',
+            valueRange: [0, null],
+            plugins: [doubleClickZoomOutPlugin],
+            underlayCallback: this.underlay
+        };
+        // set up the wavelength axis
+        let data = this.toggleframe.prop('checked') ? this.rest_spaxeldata : spaxel;
+        options = this.addDygraphWaveOptions(options);
+
+        // create the Dygraph object
+        this.webspec = new Dygraph(this.graphdiv[0], data, options);
+
+        // assign the wavelength data and bad mask regions to the Dygraph object; makes accessibile inside underlay callback
+        this.webspec.wave = this.toggleframe.prop('checked') ? this.restwave : this.obswave;
+        this.webspec.badspots = this.badspots;
+
+        // create dap line annotations and conditionally display
+        this.updateAnnotations();
     }
+
+    // This function draws highlighted masks on the Dygraph canvas
+    underlay(canvas, area, g) {
+
+        canvas.fillStyle = 'rgb(205, 92, 92, 0.5)';
+
+        if (g.badspots == null ) {
+            return;
+        }
+
+        function highlight_mask(x_start, x_end) {
+            let bottom_left = g.toDomCoords(x_start, -20);
+            let top_right = g.toDomCoords(x_end, +20);
+
+            let left = bottom_left[0];
+            let right = top_right[0];
+            canvas.fillRect(left, area.y, right - left, area.h);
+        }
+        g.badspots.forEach( x => {
+            let xstart = g.wave[x[0]];
+            let xend = g.wave[x[1]];
+            highlight_mask(xstart, xend);
+        })
+        
+    }
+
+    // set up some spaxel data arrays
+    setupSpaxel(spaxel) {
+        this._spaxeldata = spaxel;
+        this.computeRestWave();
+        this.annotations = this.daplines.map(this.createAnnotation, this);
+    } 
 
     // Dygraph Axis Formatter
     setSpectrumAxisFormatter(wave, redshift) {                
         let obs = (d, gran) => d;
-        let rest = (d, gran) => d / (1 + redshift);
+        let rest = (d, gran) => parseFloat((d / (1 + redshift)).toPrecision(5));
 
         if (wave === 'obs') {
             return obs;            
          } else if (wave === 'rest') {
             return rest;
          }
+    }
+
+    // update the Dygraph x-axis label and data with rest-frame / obs wavelength
+    addDygraphWaveOptions(oldoptions) {
+        let newopts = {};
+        if (this.toggleframe.prop('checked')) {
+            newopts = { 'file': this.rest_spaxeldata, 'xlabel': 'Rest Wavelength [Ångströms]' };
+        } else {
+            newopts = { 'file': this._spaxeldata, 'xlabel': 'Observed Wavelength [Ångströms]' };
+        }
+        let options = Object.assign(oldoptions, newopts);
+        return options;         
+    }
+
+    // Toggle the Observed/Rest Wavelength
+    toggleWavelength(event) {
+        let _this = event.data;
+        let options = {};
+        options = _this.addDygraphWaveOptions(options);
+        _this.webspec.updateOptions(options);
+        _this.updateAnnotations();
+        _this.updateMask(_this.badspots);
+    }
+
+    // Update the line annotations on Dygraph
+    updateAnnotations() {
+        if (this.togglelines.prop('checked')) {
+            this.annotations = this.daplines.map(this.createAnnotation, this);
+            this.webspec.setAnnotations(this.annotations);
+        } else {
+            this.webspec.setAnnotations([]);
+        }
+    }
+
+    // Toggle Line Display
+    toggleLines(event) {
+        let _this = event.data;
+        _this.updateAnnotations();
+    }
+
+    // create a Dygraph annotation for measured DAP lines
+    createAnnotation(x, index, arr) {
+        let [name, wave] = x.split(' ');
+        name = name.includes('-') ? name.split('-')[0] + name.split('-')[1][0] : name;
+        let owave = parseInt(wave) * (1 + this.redshift);
+        let wavedata = (this.toggleframe.prop('checked')) ? this.restwave : this.obswave;
+        let use_wave = (this.toggleframe.prop('checked')) ? wave : owave;
+        let diff = wavedata.map(y => Math.abs(y - use_wave));
+        let idx = diff.indexOf(Math.min(...diff));
+        let closest_wave = String(wavedata[idx]);
+        let annot = { 'series': "Flux", 'x': closest_wave, 'shortText': name, 'text': x, 'width': 30, 'height': 20, 'tickHeight': 40, 'tickColor': '#cd5c5c', 'cssClass': 'annotation'};
+        return annot;
+    }
+
+    // Toggle DONOTUSE Masks
+    toggleMask(event) {
+        let _this = event.data; 
+        if (_this.togglemask.prop('checked')) {
+            _this.webspec.badspots = null;
+        } else {
+            _this.webspec.badspots = _this.badspots;
+        }
+        _this.webspec.updateOptions({});
+    }
+
+    // Update the masks with new bad regions 
+    updateMask(badspots) {
+        this.badspots = badspots == null ? this.badspots : badspots;
+        this.webspec.wave = this.toggleframe.prop('checked') ? this.restwave : this.obswave;
+        this.webspec.badspots = this.togglemask.prop('checked') ? null : badspots;
+        this.webspec.updateOptions({});
     }
 
     // Update the spectrum message div for errors only
@@ -183,10 +312,15 @@ class Galaxy {
         this.specmsg.html(newmsg);
     }
 
-    // Update a DyGraph spectrum
-    updateSpaxel(spaxel, specmsg) {
+    // Update a DyGraph spectrum after clicking on a Map or the Image
+    updateSpaxel(spaxel, specmsg, badspots) {
+        this.setupSpaxel(spaxel);
         this.updateSpecMsg(specmsg);
-        this.webspec.updateOptions({'file': spaxel, 'title':specmsg});
+        let options = {'title': specmsg};
+        options = this.addDygraphWaveOptions(options);
+        this.webspec.updateOptions(options);
+        this.updateAnnotations();
+        this.updateMask(badspots);
     }
 
     // Initialize OpenLayers Map
@@ -197,6 +331,7 @@ class Galaxy {
         this.olmap.map.on('singleclick', this.getSpaxel, this);
     }
 
+    // Initialize the DAP Heatmap displays
     initHeatmap(maps) {
         let mapchildren = this.mapsdiv.children('div');
         let _this = this;
@@ -233,7 +368,7 @@ class Galaxy {
                 if (data.result.status === -1) {
                     throw new SpaxelError(`Error: ${data.result.specmsg}`);
                 }
-                this.updateSpaxel(data.result.spectra, data.result.specmsg);
+                this.updateSpaxel(data.result.spectra, data.result.specmsg, data.result.badspots);
             })
             .catch((error)=>{
                 let errmsg = (error.message === undefined) ? this.makeError('getSpaxel') : error.message;
@@ -313,11 +448,17 @@ class Galaxy {
                             throw new MapError(`Error: ${data.result.mapmsg}`);
                         }
 
+                        // extract some properties from the result
                         let image = data.result.image;
                         let spaxel = data.result.spectra;
                         let spectitle = data.result.specmsg;
                         let maps = data.result.maps;
                         let mapmsg = data.result.mapmsg;
+
+                        // add the list of DAP lines and bad mask regions to the global scope
+                        _this.daplines = data.result.daplines;
+                        _this.badspots = data.result.badspots;
+
                         // Load the Galaxy Image
                         _this.initOpenLayers(image);
                         _this.toggleload.hide();
@@ -341,6 +482,8 @@ class Galaxy {
     initFlagPopovers() {
         // DRP Quality Popovers
         this.qualpop.popover({html:true,content:$('#list_drp3quality').html()});
+        // DAP Map Quality Popovers
+        this.mapqualpop.popover({html:true,content:$('#list_dapquality').html()});
         // MaNGA Target Popovers
         $.each(this.targpops, function(index, value) {
             // get id of flag link